data type fix

+++ Models/Developer.cs
@@ -4,7 +4,7 @@ namespace EmployeesSalaries.Models
     public class Developer: Employee, IManaged<LeadDeveloper>
     {
         public LeadDeveloper Manager { get; set; }
-        public Developer(int id, int basicSalary): base (id, basicSalary)
+        public Developer(int id, double basicSalary): base (id, basicSalary)
         {
         }
